Убрал лишние папки, начал настраивать докер
