+++ .gitignore
@@ -1,0 +1,4 @@
+/node_modules
+/build
+.env*
+package-lock.json
