+++ README.md
@@ -1,0 +1,9 @@
+# Фронтенд часть приложения инвентаризации колледжа
+
+## Настройка
+
+1. Подготовить `.env` файл
+
+```md
+    REACT_APP_PORT=3000
+```
