--- client/src/components/addAudience/AddAudience.scss
@@ -1,27 +1,0 @@
-// select{
-//     padding: 20px;
-// }
-
-.wrapperAddAudience{
-    width: 400px;
-        height: max-content;
-        border: 1px solid #C82647;
-        border-radius: 5px;
-        padding: 20px 20px;
-        &>input, select{
-            width: 100%;
-            margin-bottom: 20px;
-            padding: 5px;
-        }
-        &>button{
-            background-color: white;
-            border-color: #C82647;
-            width: 100%;
-            padding: 5px;
-        }
-        // &>select{
-        //     width: 100%;
-        //     padding: 5px;
-        //     margin-bottom: 20px;
-        // }
-}
