--- client/src/components/addObject/AddObject.tsx
@@ -1,15 +1,0 @@
-// import React from 'react'
-
-const AddObject = () => {
-  return (
-    <div>
-        <select>
-            <option>asdasdasdasd</option>
-        </select>
-        <input placeholder=''></input>
-        
-    </div>
-  )
-}
-
-export default AddObject
