--- client/src/components/footer/Footer.scss
@@ -1,4 +1,0 @@
-footer{
-    height: 150px;
-    background-color: #C82647;
-}
