--- client/src/components/footer/Footer.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import "./Footer.scss"
-
-const Footer = () => {
-  return (
-    <footer></footer>
-  )
-}
-
-export default Footer
