--- client/src/pages/about/About.tsx
@@ -1,7 +1,0 @@
-const About = () => {
-  return (
-    <main>About</main>
-  )
-}
-
-export default About
