--- client/src/pages/organization/Organization.tsx
@@ -1,11 +1,0 @@
-import AddOrganization from '../../components/addOrganization/AddOrganization'
-
-const Organization = () => {
-  return (
-    <main>
-      <AddOrganization/>
-    </main>
-  )
-}
-
-export default Organization
