--- client/src/pages/profile/Profile.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Profile = () => {
-  return (
-    <main>
-      
-    </main>
-  )
-}
-
-export default Profile
