+++ docker-compose.yml
@@ -1,0 +1,9 @@
+version: "3"
+services:
+  qr-inventory:
+    container_name: qr-inventory
+    build: .
+    environment:
+      env_file: .env
+    ports:
+      - "80:80"
