+++ src/components/addOrganization/AddOrganization.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+
+const AddOrganization = () => {
+  return (
+    <div>
+        <input placeholder=''></input>
+        <select>
+            <option>asdasdasdasd</option>
+        </select>
+    </div>
+  )
+}
+
+export default AddOrganization
