+++ src/components/footer/Footer.scss
@@ -1,0 +1,4 @@
+footer{
+    height: 150px;
+    background-color: #C82647;
+}
