+++ src/components/navbar/Navbar.scss
@@ -1,0 +1,28 @@
+nav{
+    height: 60px;
+    background-color: #C82647;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 0 30px;
+    &>a>img{
+        width: 10%;
+    }
+    &>ul{
+        display: flex;
+        color: white;
+        list-style: none;
+        align-items:center;
+        justify-content: space-between;
+        &>a>img{
+            width: 70%;
+        }
+        &>li{
+            margin-right:40px;
+&>a{
+    text-decoration: none;
+    color:white;
+}
+        }
+    }
+}
