+++ src/pages/about/About.tsx
@@ -1,0 +1,7 @@
+const About = () => {
+  return (
+    <main>About</main>
+  )
+}
+
+export default About
