+++ src/pages/object/Object.tsx
@@ -1,0 +1,11 @@
+import AddObject from '../../components/addObject/AddObject'
+
+const Object = () => {
+  return (
+    <main>
+      <AddObject/>
+    </main>
+  )
+}
+
+export default Object
