+++ src/pages/organization/Organization.tsx
@@ -1,0 +1,11 @@
+import AddOrganization from '../../components/addOrganization/AddOrganization'
+
+const Organization = () => {
+  return (
+    <main>
+      <AddOrganization/>
+    </main>
+  )
+}
+
+export default Organization
