+++ src/pages/profile/Profile.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Profile = () => {
+  return (
+    <main>
+      
+    </main>
+  )
+}
+
+export default Profile
